Updated read mpu gyro example

+++ examples/read_mpu9250_gyro.cc
@@ -8,12 +8,12 @@
 int main() {
   I2cBus i2c_1(1);
 
-  int addr = 0x68;  // MPU9250 address
-  int gyro_xout_h = 0x43;  // starting MPU register address from which to read
+  uint16_t addr = 0x68;  // MPU9250 address
+  uint8_t gyro_xout_h = 0x43;  // starting MPU register address from which to read
   uint8_t gyro[6] = {0};
-  int gyro_x = 0;
-  int gyro_y = 0;
-  int gyro_z = 0;
+  uint16_t gyro_x = 0;
+  uint16_t gyro_y = 0;
+  uint16_t gyro_z = 0;
 
   // Read 6 register values starting from gyro_xout_h
   i2c_1.ReadFromMemInto(addr, gyro_xout_h, 6, gyro);
